% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kastore_version}
\alias{kastore_version}
\title{Report the version of installed kastore C API}
\usage{
kastore_version()
}
\value{
A named vector with three elements \code{major}, \code{minor}, and
  \code{patch}.
}
\description{
Report the version of installed kastore C API
}
\details{
The version is stored in the installed header \code{kastore.h}.
}
\examples{
kastore_version()
}
