% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{treeseq_num_nodes}
\alias{treeseq_num_nodes}
\title{Get number of nodes in a tree sequence}
\usage{
treeseq_num_nodes(ts)
}
\arguments{
\item{ts}{an external pointer to a \code{tsk_treeseq_t} object.}
}
\value{
An integer specifying the number of nodes in the tree sequence.
}
\description{
Get number of nodes in a tree sequence
}
\examples{
ts_file <- system.file("examples", "test.trees", package = "tskitr")
ts <- treeseq_load(ts_file)
treeseq_num_nodes(ts)
n <- treeseq_num_nodes(ts)
n
is(n)
try(treeseq_num_nodes())
try(treeseq_num_nodes(nonexistent_ts))
}
