% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-TreeSequence.R
\name{TreeSequence-read-alias}
\alias{TreeSequence-read-alias}
\title{Create a \code{\link{TreeSequence}} from a file.}
\arguments{
\item{file}{see \code{\link{ts_load}()}}

\item{options}{see \code{\link{ts_load}()}}
}
\description{
Alias for \code{TreeSequence$new(file, options)}
  See \code{\link{ts_load}()} for details and examples.
}
