% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/RcppTskit-package.R
\name{ts_dump_ptr}
\alias{ts_dump_ptr}
\alias{ts_write_ptr}
\title{Write a tree sequence to a file}
\usage{
ts_dump_ptr(ts, file, options = 0L)

ts_write_ptr(ts, file, options = 0L)
}
\arguments{
\item{ts}{tree sequence as a \code{\link{TreeSequence}} object or
an external pointer to tree sequence as a \code{tsk_treeseq_t} object.}

\item{file}{a string specifying the full path of the tree sequence file}

\item{options}{integer bitwise options (see details).}
}
\value{
No return value; called for side effects.
}
\description{
Write a tree sequence to a file
}
\details{
This function calls the tskit C API: \code{tsk_treeseq_dump()}
  \url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_dump}.
}
\section{Functions}{
\itemize{
\item \code{ts_write_ptr()}: Alias for \code{ts_dump_ptr()}

}}
\examples{
ts_file <- system.file("examples/test.trees", package = "RcppTskit")

# TreeSequence class object
ts <- ts_load(ts_file)
dump_file <- tempfile()
ts$dump(dump_file)

# External pointer object
ts_ptr <- ts$pointer
ts_dump_ptr(ts_ptr, dump_file)
\dontshow{on.exit(file.remove(dump_file)) # clean up}
}
