% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/RcppTskit-package.R
\name{ts_load_ptr}
\alias{ts_load_ptr}
\alias{ts_load}
\alias{ts_read}
\alias{ts_read_ptr}
\title{Load a tree sequence from a file}
\usage{
ts_load_ptr(file, options = 0L)

ts_load(file, options = 0L)

ts_read(file, options = 0L)

ts_read_ptr(file, options = 0L)
}
\arguments{
\item{file}{a string specifying the full path of the tree sequence file.}

\item{options}{integer bitwise options (see details).}
}
\value{
\code{ts_load()} returns a \code{\link{TreeSequence}} object.
 \code{ts_load_ptr()} returns external pointer to tree sequence as a
 \code{tsk_treeseq_t} object.
}
\description{
Load a tree sequence from a file
}
\details{
This function calls tskit C API: \code{tsk_treeseq_load()}
  \url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_load}.
}
\section{Functions}{
\itemize{
\item \code{ts_load()}: Load a tree sequence from a file

\item \code{ts_read()}: Alias for \code{ts_load()}

\item \code{ts_read_ptr()}: Alias for \code{ts_load_ptr()}

}}
\examples{
ts_file <- system.file("examples/test.trees", package = "RcppTskit")

# TreeSequence class object
ts <- ts_load(ts_file)
is(ts)
ts
ts$num_nodes()

# External pointer object
ts_ptr <- ts_load_ptr(ts_file)
is(ts_ptr)
ts_ptr
ts_num_nodes_ptr(ts_ptr)

# Also
ts <- TreeSequence$new(file = ts_file); is(ts)
ts <- TreeSequence$new(pointer = ts_ptr); is(ts)
}
