% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ts_num_provenances_ptr}
\alias{ts_num_provenances_ptr}
\alias{ts_num_populations_ptr}
\alias{ts_num_migrations_ptr}
\alias{ts_num_individuals_ptr}
\alias{ts_num_samples_ptr}
\alias{ts_num_nodes_ptr}
\alias{ts_num_edges_ptr}
\alias{ts_num_trees_ptr}
\alias{ts_num_sites_ptr}
\alias{ts_num_mutations_ptr}
\alias{ts_sequence_length_ptr}
\alias{ts_time_units_ptr}
\alias{ts_summary_ptr}
\title{Summary of properties and number of records in a tree sequence}
\usage{
ts_num_provenances_ptr(ts)

ts_num_populations_ptr(ts)

ts_num_migrations_ptr(ts)

ts_num_individuals_ptr(ts)

ts_num_samples_ptr(ts)

ts_num_nodes_ptr(ts)

ts_num_edges_ptr(ts)

ts_num_trees_ptr(ts)

ts_num_sites_ptr(ts)

ts_num_mutations_ptr(ts)

ts_sequence_length_ptr(ts)

ts_time_units_ptr(ts)

ts_summary_ptr(ts)
}
\arguments{
\item{ts}{tree sequence as a \code{\link{TreeSequence}} object or
an external pointer to tree sequence as a \code{tsk_treeseq_t} object.}
}
\value{
\code{ts_summary()} and \code{ts_summary_ptr()} return
  a named list with the number/value for all items,
  while other functions \code{ts_num_x()} and \code{ts_num_x_ptr()} etc.
  return the number/value of each item.
}
\description{
Summary of properties and number of records in a tree sequence
}
\details{
These functions return the summary of properties and number of
  records in a tree sequence, by calling tskit C API:
  \code{tsk_treeseq_get_num_provenances()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_provenances},
  \code{tsk_treeseq_get_num_populations()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_populations},
  \code{tsk_treeseq_get_num_migrations()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_migrations},
  \code{tsk_treeseq_get_num_individuals()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_individuals},
  \code{tsk_treeseq_get_num_samples()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_samples},
  \code{tsk_treeseq_get_num_nodes()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_nodes},
  \code{tsk_treeseq_get_num_edges()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_edges},
  \code{tsk_treeseq_get_num_trees()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_trees},
  \code{tsk_treeseq_get_num_sites()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_sites},
  \code{tsk_treeseq_get_num_mutations()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_mutations},
  \code{tsk_treeseq_get_sequence_length()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_sequence_length},
  and
  \code{tsk_treeseq_get_time_units()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_time_units}.
}
\section{Functions}{
\itemize{
\item \code{ts_num_provenances_ptr()}: Get the number of provenances in a tree sequence

\item \code{ts_num_populations_ptr()}: Get the number of populations in a tree sequence

\item \code{ts_num_migrations_ptr()}: Get the number of migrations in a tree sequence

\item \code{ts_num_individuals_ptr()}: Get the number of individuals in a tree sequence

\item \code{ts_num_samples_ptr()}: Get the number of samples (of nodes) in a tree
sequence

\item \code{ts_num_nodes_ptr()}: Get the number of nodes in a tree sequence

\item \code{ts_num_edges_ptr()}: Get the number of edges in a tree sequence

\item \code{ts_num_trees_ptr()}: Get the number of trees in a tree sequence

\item \code{ts_num_sites_ptr()}: Get the number of sites in a tree sequence

\item \code{ts_num_mutations_ptr()}: Get the number of mutations in a tree sequence

\item \code{ts_sequence_length_ptr()}: Get the sequence length

\item \code{ts_time_units_ptr()}: Get the time units string

}}
\examples{
ts_file <- system.file("examples/test.trees", package = "RcppTskit")

# TreeSequence class object
ts <- ts_load(ts_file)
ts$print()
ts$summary()
ts$num_provenances()
ts$num_populations()
ts$num_migrations()
ts$num_individuals()
ts$num_samples()
ts$num_nodes()
ts$num_edges()
ts$num_trees()
ts$num_sites()
ts$num_mutations()
ts$sequence_length()
ts$time_units()

# External pointer object
ts_ptr <- ts$pointer
ts_summary_ptr(ts_ptr)
ts_num_provenances_ptr(ts_ptr)
ts_num_populations_ptr(ts_ptr)
ts_num_migrations_ptr(ts_ptr)
ts_num_individuals_ptr(ts_ptr)
ts_num_samples_ptr(ts_ptr)
ts_num_nodes_ptr(ts_ptr)
ts_num_edges_ptr(ts_ptr)
ts_num_trees_ptr(ts_ptr)
ts_num_sites_ptr(ts_ptr)
ts_num_mutations_ptr(ts_ptr)
ts_sequence_length_ptr(ts_ptr)
ts_time_units_ptr(ts_ptr)
}
