% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tskit_version}
\alias{tskit_version}
\title{Report the version of installed tskit C API}
\usage{
tskit_version()
}
\value{
A named vector with three elements \code{major}, \code{minor}, and
  \code{patch}.
}
\description{
Report the version of installed tskit C API
}
\details{
The version is defined in the installed header \code{tskit/core.h}.
}
\examples{
tskit_version()
}
