% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppTskit-package.R
\docType{package}
\name{RcppTskit-package}
\alias{RcppTskit}
\alias{RcppTskit-package}
\title{RcppTskit: R access to the tskit C API}
\description{
Tskit enables performant storage, manipulation, and analysis of
ancestral recombination graphs (ARGs) using succinct tree sequence encoding.
See https://tskit.dev for project news, documentation, and tutorials.
Tskit provides Python, C, and Rust APIs. The Python API can be called from R
via the `reticulate` R package to seamlessly load and analyse a tree sequence
as described at https://tskit.dev/tutorials/tskitr.html.
`RcppTskit` provides R access to the `tskit` C API for use cases where the
`reticulate` approach is not optimal. For example, for high-performance
and low-level work with tree sequences. Currently, `RcppTskit` provides a very
limited number of R functions due to the availability of extensive Python API
and the `reticulate` approach.
}
\examples{
\dontshow{# Providing the examples here so we test them via R CMD check}
# Here are examples showcasing what you can do with RcppTskit

# 1) Load a tree sequence into R and summarise it
# Load a tree sequence
ts_file <- system.file("examples/test.trees", package = "RcppTskit")
ts <- ts_load(ts_file)

# Print summary of the tree sequence
ts$num_individuals()
ts

# 2) Pass tree sequence between R and reticulate or standard Python

# Tree sequence in R
ts_file <- system.file("examples/test.trees", package = "RcppTskit")
ts <- ts_load(ts_file)

# If you have a tree sequence in R and you want to use tskit Python API,
# you can write it to disk and load it into reticulate Python
ts_py <- ts$r_to_py()
# ... continue in reticulate Python ...
ts_py$num_individuals # 80
ts2_py = ts_py$simplify(samples = c(0L, 1L, 2L, 3L))
ts2_py$num_individuals # 2
# ... and to bring it back to R ...
ts2 <- ts_py_to_r(ts2_py)
ts2$num_individuals() # 2

# If you prefer standard (non-reticulate) Python, use approach:
ts_file <- tempfile()
print(ts_file)
ts$dump(file = ts_file)
# ... continue in standard Python ...
# import tskit
# ts = tskit.load("insert_ts_file_path_here")
# ts.num_individuals # 80
# ts2 = ts.simplify(samples = [0, 1, 2, 3])
# ts2.num_individuals # 2
# ts2.dump("insert_ts_file_path_here")
# ... and to bring it back to R ...
ts2 <- ts_load(ts_file)
ts$num_individuals() # 2 (if you have ran the above Python code)
\dontshow{on.exit(file.remove(ts_file)) # clean up}

# 3) Call tskit C API in C++ code in R session or script
library(Rcpp)
# Write and compile a C++ function
codeString <- '
  #include <tskit.h>
  int ts_num_individuals(SEXP ts) {
    Rcpp::XPtr<tsk_treeseq_t> ts_xptr(ts);
    return (int) tsk_treeseq_get_num_individuals(ts_xptr);
  }'
ts_num_individuals2 <- Rcpp::cppFunction(code=codeString,
                                         depends="RcppTskit",
                                         plugins="RcppTskit")
# We must specify both the `depends` and `plugins` arguments!

# Load a tree sequence
ts_file <- system.file("examples/test.trees", package="RcppTskit")
ts <- ts_load(ts_file)

# Apply the compiled function
ts_num_individuals2(ts$pointer)

# An identical RcppTskit implementation
ts$num_individuals()
ts_num_individuals_ptr(ts$pointer)

# 4) Call `tskit` C API in C++ code in another R package
# TODO: Show vignette here
#       https://github.com/HighlanderLab/RcppTskit/issues/10
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/highlanderlab/RcppTskit}
}

}
\author{
\strong{Maintainer}: Gregor Gorjanc \email{gregor.gorjanc@gmail.com} (\href{https://orcid.org/0000-0001-8008-2787}{ORCID})

}
\keyword{internal}
