% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppTskit-package.R
\docType{package}
\name{RcppTskit-package}
\alias{RcppTskit}
\alias{RcppTskit-package}
\title{RcppTskit: R access to the tskit C API}
\description{
Tskit enables performant storage, manipulation, and analysis of
ancestral recombination graphs (ARGs) using succinct tree sequence encoding.
See https://tskit.dev for project news, documentation, and tutorials.
Tskit provides Python, C, and Rust APIs. The Python API can be called from R
via the `reticulate` R package to seamlessly load and analyse a tree sequence
as described at https://tskit.dev/tutorials/tskitr.html.
`RcppTskit` provides R access to the `tskit` C API for use cases where the
`reticulate` option is not optimal. For example, for high-performance
and low-level work with tree sequences. Currently, `RcppTskit` provides a
limited number of R functions due to the availability of extensive Python API
and the `reticulate` option.
}
\examples{
vignette(package="RcppTskit")
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/highlanderlab/RcppTskit}
}

}
\author{
\strong{Maintainer}: Gregor Gorjanc \email{gregor.gorjanc@gmail.com} (\href{https://orcid.org/0000-0001-8008-2787}{ORCID})

}
\keyword{internal}
