% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-TableCollection.R
\name{TableCollection}
\alias{TableCollection}
\title{Table collection R6 class (TableCollection)}
\description{
An \code{R6} class holding an external pointer to
a table collection object. As an \code{R6} class, method-calling looks Pythonic
and hence resembles the \code{tskit Python} API. Since the class only
holds the pointer, it is lightweight. Currently there is a limited set of
\code{R} methods for working with the table collection object.
}
\examples{

## ------------------------------------------------
## Method `TableCollection$new`
## ------------------------------------------------

ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- TableCollection$new(file = ts_file)
is(tc)
tc

## ------------------------------------------------
## Method `TableCollection$dump`
## ------------------------------------------------

ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- TableCollection$new(file = ts_file)
dump_file <- tempfile()
tc$dump(dump_file)
tc$write(dump_file) # alias
\dontshow{file.remove(dump_file)}

## ------------------------------------------------
## Method `TableCollection$tree_sequence`
## ------------------------------------------------

ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- TableCollection$new(file = ts_file)
ts <- tc$tree_sequence()
is(ts)

## ------------------------------------------------
## Method `TableCollection$num_provenances`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_provenances()

## ------------------------------------------------
## Method `TableCollection$num_populations`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_populations()

## ------------------------------------------------
## Method `TableCollection$num_migrations`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_migrations()

## ------------------------------------------------
## Method `TableCollection$num_individuals`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_individuals()

## ------------------------------------------------
## Method `TableCollection$individual_table_add_row`
## ------------------------------------------------

ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(ts_file)
n_before <- tc$num_individuals()
new_id <- tc$individual_table_add_row()
new_id <- tc$individual_table_add_row(location = c(5, 8))
new_id <- tc$individual_table_add_row(flags = 0L)
new_id <- tc$individual_table_add_row(parents = c(0L, 2L))
new_id <- tc$individual_table_add_row(metadata = "abc")
new_id <- tc$individual_table_add_row(metadata = charToRaw("cba"))
n_after <- tc$num_individuals()

## ------------------------------------------------
## Method `TableCollection$num_nodes`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_nodes()

## ------------------------------------------------
## Method `TableCollection$node_table_add_row`
## ------------------------------------------------

ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(ts_file)
n_before <- tc$num_nodes()
new_id <- tc$node_table_add_row()
new_id <- tc$node_table_add_row(time = 2.5)
new_id <- tc$node_table_add_row(flags = 1L, time = 3.5, population = 0L)
new_id <- tc$node_table_add_row(flags = 1L, time = 4.5, individual = 0L)
new_id <- tc$node_table_add_row(metadata = "abc")
new_id <- tc$node_table_add_row(metadata = charToRaw("cba"))
n_after <- tc$num_nodes()

## ------------------------------------------------
## Method `TableCollection$num_edges`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_edges()

## ------------------------------------------------
## Method `TableCollection$edge_table_add_row`
## ------------------------------------------------

ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(ts_file)
parent <- 0L
child <- 1L
n_before <- tc$num_edges()
new_id <- tc$edge_table_add_row(
  left = 0, right = 1, parent = parent, child = child
)
new_id <- tc$edge_table_add_row(
  left = 1, right = 2, parent = parent, child = child, metadata = "abc"
)
new_id <- tc$edge_table_add_row(
  left = 2, right = 3, parent = parent, child = child, metadata = charToRaw("cba")
)
n_after <- tc$num_edges()

## ------------------------------------------------
## Method `TableCollection$num_sites`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_sites()

## ------------------------------------------------
## Method `TableCollection$site_table_add_row`
## ------------------------------------------------

ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(ts_file)
n_before <- tc$num_sites()
new_id <- tc$site_table_add_row(position = 0.5, ancestral_state = "A")
new_id <- tc$site_table_add_row(position = 1.5, ancestral_state = charToRaw("G"))
new_id <- tc$site_table_add_row(position = 2.5, ancestral_state = "T", metadata = "abc")
n_after <- tc$num_sites()

## ------------------------------------------------
## Method `TableCollection$num_mutations`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_mutations()

## ------------------------------------------------
## Method `TableCollection$mutation_table_add_row`
## ------------------------------------------------

ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(ts_file)
n_before <- tc$num_mutations()
new_id <- tc$mutation_table_add_row(site = 0L, node = 0L, derived_state = "T")
new_id <- tc$mutation_table_add_row(
  site = 0L,
  node = 0L,
  parent = -1L,
  time = 1.5,
  derived_state = charToRaw("C"),
  metadata = "abc"
)
n_after <- tc$num_mutations()

## ------------------------------------------------
## Method `TableCollection$sequence_length`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$sequence_length()

## ------------------------------------------------
## Method `TableCollection$time_units`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$time_units()

## ------------------------------------------------
## Method `TableCollection$has_index`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$has_index()

## ------------------------------------------------
## Method `TableCollection$build_index`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$has_index()
tc$drop_index()
tc$has_index()
tc$build_index()
tc$has_index()

## ------------------------------------------------
## Method `TableCollection$drop_index`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$has_index()
tc$drop_index()
tc$has_index()

## ------------------------------------------------
## Method `TableCollection$has_reference_sequence`
## ------------------------------------------------

tc_file1 <- system.file("examples/test.trees", package = "RcppTskit")
tc_file2 <- system.file("examples/test_with_ref_seq.trees", package = "RcppTskit")
tc1 <- tc_load(tc_file1)
tc1$has_reference_sequence()
tc2 <- tc_load(tc_file2)
tc2$has_reference_sequence()

## ------------------------------------------------
## Method `TableCollection$file_uuid`
## ------------------------------------------------

tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$file_uuid()

## ------------------------------------------------
## Method `TableCollection$r_to_py`
## ------------------------------------------------

\dontrun{
  ts_file <- system.file("examples/test.trees", package = "RcppTskit")
  tc_r <- tc_load(ts_file)
  is(tc_r)
  tc_r$print()

  # Transfer the table collection to reticulate Python and use tskit Python API
  tskit <- get_tskit_py()
  if (check_tskit_py(tskit)) {
    tc_py <- tc_r$r_to_py()
    is(tc_py)
    tmp <- tc_py$simplify(samples = c(0L, 1L, 2L, 3L))
    tmp
    tc_py$individuals$num_rows # 2
    tc_py$nodes$num_rows # 8
    tc_py$nodes$time # 0.0 ... 5.0093910
  }
}

## ------------------------------------------------
## Method `TableCollection$print`
## ------------------------------------------------

ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(file = ts_file)
tc$print()
tc
}
\seealso{
\code{\link{tc_py_to_r}}, \code{\link{tc_load}}, and
  \code{\link[=TableCollection]{TableCollection$dump}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{xptr}}{external pointer to the table collection}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TableCollection-new}{\code{TableCollection$new()}}
\item \href{#method-TableCollection-dump}{\code{TableCollection$dump()}}
\item \href{#method-TableCollection-write}{\code{TableCollection$write()}}
\item \href{#method-TableCollection-tree_sequence}{\code{TableCollection$tree_sequence()}}
\item \href{#method-TableCollection-num_provenances}{\code{TableCollection$num_provenances()}}
\item \href{#method-TableCollection-num_populations}{\code{TableCollection$num_populations()}}
\item \href{#method-TableCollection-num_migrations}{\code{TableCollection$num_migrations()}}
\item \href{#method-TableCollection-num_individuals}{\code{TableCollection$num_individuals()}}
\item \href{#method-TableCollection-individual_table_add_row}{\code{TableCollection$individual_table_add_row()}}
\item \href{#method-TableCollection-num_nodes}{\code{TableCollection$num_nodes()}}
\item \href{#method-TableCollection-node_table_add_row}{\code{TableCollection$node_table_add_row()}}
\item \href{#method-TableCollection-num_edges}{\code{TableCollection$num_edges()}}
\item \href{#method-TableCollection-edge_table_add_row}{\code{TableCollection$edge_table_add_row()}}
\item \href{#method-TableCollection-num_sites}{\code{TableCollection$num_sites()}}
\item \href{#method-TableCollection-site_table_add_row}{\code{TableCollection$site_table_add_row()}}
\item \href{#method-TableCollection-num_mutations}{\code{TableCollection$num_mutations()}}
\item \href{#method-TableCollection-mutation_table_add_row}{\code{TableCollection$mutation_table_add_row()}}
\item \href{#method-TableCollection-sequence_length}{\code{TableCollection$sequence_length()}}
\item \href{#method-TableCollection-time_units}{\code{TableCollection$time_units()}}
\item \href{#method-TableCollection-has_index}{\code{TableCollection$has_index()}}
\item \href{#method-TableCollection-build_index}{\code{TableCollection$build_index()}}
\item \href{#method-TableCollection-drop_index}{\code{TableCollection$drop_index()}}
\item \href{#method-TableCollection-has_reference_sequence}{\code{TableCollection$has_reference_sequence()}}
\item \href{#method-TableCollection-file_uuid}{\code{TableCollection$file_uuid()}}
\item \href{#method-TableCollection-r_to_py}{\code{TableCollection$r_to_py()}}
\item \href{#method-TableCollection-print}{\code{TableCollection$print()}}
\item \href{#method-TableCollection-clone}{\code{TableCollection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-new"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{\link{TableCollection}} from a file or an external pointer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$new(
  file,
  skip_tables = FALSE,
  skip_reference_sequence = FALSE,
  xptr = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{a string specifying the full path of the tree sequence file.}

\item{\code{skip_tables}}{logical; if \code{TRUE}, load only non-table information.}

\item{\code{skip_reference_sequence}}{logical; if \code{TRUE}, skip loading
reference genome sequence information.}

\item{\code{xptr}}{an external pointer (\code{externalptr}) to a table collection.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See the \code{tskit Python} equivalent at
  \url{https://github.com/tskit-dev/tskit/blob/dc394d72d121c99c6dcad88f7a4873880924dd72/python/tskit/tables.py#L3463}.
  TODO: Update URL to TableCollection.load() method #104
        https://github.com/HighlanderLab/RcppTskit/issues/104
}

\subsection{Returns}{
A \code{\link{TableCollection}} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- TableCollection$new(file = ts_file)
is(tc)
tc
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-dump"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-dump}{}}}
\subsection{Method \code{dump()}}{
Write a table collection to a file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$dump(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{a string specifying the full path of the tree sequence file.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See the \code{tskit Python} equivalent at
  \url{https://tskit.dev/tskit/docs/latest/python-api.html#tskit.TableCollection.dump}.
}

\subsection{Returns}{
No return value; called for side effects.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- TableCollection$new(file = ts_file)
dump_file <- tempfile()
tc$dump(dump_file)
tc$write(dump_file) # alias
\dontshow{file.remove(dump_file)}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-write"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-write}{}}}
\subsection{Method \code{write()}}{
Alias for \code{\link[=TableCollection]{TableCollection$dump}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$write(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{see \code{\link[=TableCollection]{TableCollection$dump}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-tree_sequence"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-tree_sequence}{}}}
\subsection{Method \code{tree_sequence()}}{
Create a \code{\link{TreeSequence}} from this table collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$tree_sequence()}\if{html}{\out{</div>}}
}

\subsection{Details}{
See the \code{tskit Python} equivalent at
  \url{https://tskit.dev/tskit/docs/latest/python-api.html#tskit.TableCollection.tree_sequence}.
}

\subsection{Returns}{
A \code{\link{TreeSequence}} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- TableCollection$new(file = ts_file)
ts <- tc$tree_sequence()
is(ts)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-num_provenances"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-num_provenances}{}}}
\subsection{Method \code{num_provenances()}}{
Get the number of provenances in a table collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$num_provenances()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A signed 64 bit integer \code{bit64::integer64}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_provenances()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-num_populations"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-num_populations}{}}}
\subsection{Method \code{num_populations()}}{
Get the number of populations in a table collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$num_populations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A signed 64 bit integer \code{bit64::integer64}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_populations()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-num_migrations"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-num_migrations}{}}}
\subsection{Method \code{num_migrations()}}{
Get the number of migrations in a table collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$num_migrations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A signed 64 bit integer \code{bit64::integer64}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_migrations()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-num_individuals"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-num_individuals}{}}}
\subsection{Method \code{num_individuals()}}{
Get the number of individuals in a table collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$num_individuals()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A signed 64 bit integer \code{bit64::integer64}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_individuals()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-individual_table_add_row"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-individual_table_add_row}{}}}
\subsection{Method \code{individual_table_add_row()}}{
Add a row to the individuals table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$individual_table_add_row(
  flags = 0L,
  location = NULL,
  parents = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flags}}{integer flags for the new individual.}

\item{\code{location}}{numeric vector with the location of the new individual.}

\item{\code{parents}}{integer vector with parent individual IDs (0-based).}

\item{\code{metadata}}{for the new individual; accepts \code{NULL},
a raw vector, or a character of length 1.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See the \code{tskit Python} equivalent at
  \url{https://tskit.dev/tskit/docs/stable/python-api.html#tskit.IndividualTable.add_row}.
  The function casts inputs to the expected class.
}

\subsection{Returns}{
Integer row ID (0-based) of the newly added individual.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(ts_file)
n_before <- tc$num_individuals()
new_id <- tc$individual_table_add_row()
new_id <- tc$individual_table_add_row(location = c(5, 8))
new_id <- tc$individual_table_add_row(flags = 0L)
new_id <- tc$individual_table_add_row(parents = c(0L, 2L))
new_id <- tc$individual_table_add_row(metadata = "abc")
new_id <- tc$individual_table_add_row(metadata = charToRaw("cba"))
n_after <- tc$num_individuals()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-num_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-num_nodes}{}}}
\subsection{Method \code{num_nodes()}}{
Get the number of nodes in a table collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$num_nodes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A signed 64 bit integer \code{bit64::integer64}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_nodes()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-node_table_add_row"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-node_table_add_row}{}}}
\subsection{Method \code{node_table_add_row()}}{
Add a row to the nodes table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$node_table_add_row(
  flags = 0L,
  time = 0,
  population = -1L,
  individual = -1L,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flags}}{integer flags for the new node.}

\item{\code{time}}{numeric time value for the new node.}

\item{\code{population}}{integer population row ID (0-based, or \code{-1});
\code{NULL} maps to \code{-1}.}

\item{\code{individual}}{integer individual row ID (0-based, or \code{-1});
\code{NULL} maps to \code{-1}.}

\item{\code{metadata}}{for the new node; accepts \code{NULL},
a raw vector, or a character of length 1.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See the \code{tskit Python} equivalent at
  \url{https://tskit.dev/tskit/docs/stable/python-api.html#tskit.NodeTable.add_row}.
  The function casts inputs to the expected class. For convenience,
  \code{population = NULL} and \code{individual = NULL} are mapped to
  \code{-1} (\code{TSK_NULL}).
}

\subsection{Returns}{
Integer row ID (0-based) of the newly added node.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(ts_file)
n_before <- tc$num_nodes()
new_id <- tc$node_table_add_row()
new_id <- tc$node_table_add_row(time = 2.5)
new_id <- tc$node_table_add_row(flags = 1L, time = 3.5, population = 0L)
new_id <- tc$node_table_add_row(flags = 1L, time = 4.5, individual = 0L)
new_id <- tc$node_table_add_row(metadata = "abc")
new_id <- tc$node_table_add_row(metadata = charToRaw("cba"))
n_after <- tc$num_nodes()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-num_edges"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-num_edges}{}}}
\subsection{Method \code{num_edges()}}{
Get the number of edges in a table collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$num_edges()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A signed 64 bit integer \code{bit64::integer64}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_edges()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-edge_table_add_row"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-edge_table_add_row}{}}}
\subsection{Method \code{edge_table_add_row()}}{
Add a row to the edges table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$edge_table_add_row(left, right, parent, child, metadata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{left}}{numeric scalar left coordinate for the new edge.}

\item{\code{right}}{numeric scalar right coordinate for the new edge.}

\item{\code{parent}}{integer scalar parent node row ID (0-based).}

\item{\code{child}}{integer scalar child node row ID (0-based).}

\item{\code{metadata}}{for the new edge; accepts \code{NULL},
a raw vector, or a character of length 1.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See the \code{tskit Python} equivalent at
  \url{https://tskit.dev/tskit/docs/stable/python-api.html#tskit.EdgeTable.add_row}.
  The function casts inputs to the expected class. Inputs are validated:
  \code{left} and \code{right} must be finite numeric scalars with
  \code{left < right}, and \code{parent} and \code{child} must be
  non-\code{NA} integer scalars.
}

\subsection{Returns}{
Integer row ID (0-based) of the newly added edge.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(ts_file)
parent <- 0L
child <- 1L
n_before <- tc$num_edges()
new_id <- tc$edge_table_add_row(
  left = 0, right = 1, parent = parent, child = child
)
new_id <- tc$edge_table_add_row(
  left = 1, right = 2, parent = parent, child = child, metadata = "abc"
)
new_id <- tc$edge_table_add_row(
  left = 2, right = 3, parent = parent, child = child, metadata = charToRaw("cba")
)
n_after <- tc$num_edges()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-num_sites"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-num_sites}{}}}
\subsection{Method \code{num_sites()}}{
Get the number of sites in a table collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$num_sites()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A signed 64 bit integer \code{bit64::integer64}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_sites()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-site_table_add_row"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-site_table_add_row}{}}}
\subsection{Method \code{site_table_add_row()}}{
Add a row to the sites table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$site_table_add_row(
  position,
  ancestral_state = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{numeric scalar site position.}

\item{\code{ancestral_state}}{for the new site; accepts \code{NULL},
a raw vector, or a character of length 1.}

\item{\code{metadata}}{for the new site; accepts \code{NULL},
a raw vector, or a character of length 1.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See the \code{tskit Python} equivalent at
  \url{https://tskit.dev/tskit/docs/stable/python-api.html#tskit.SiteTable.add_row}.
  The function casts inputs to the expected class. \code{position}
  must be a non-\code{NA} finite numeric scalar.
}

\subsection{Returns}{
Integer row ID (0-based) of the newly added site.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(ts_file)
n_before <- tc$num_sites()
new_id <- tc$site_table_add_row(position = 0.5, ancestral_state = "A")
new_id <- tc$site_table_add_row(position = 1.5, ancestral_state = charToRaw("G"))
new_id <- tc$site_table_add_row(position = 2.5, ancestral_state = "T", metadata = "abc")
n_after <- tc$num_sites()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-num_mutations"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-num_mutations}{}}}
\subsection{Method \code{num_mutations()}}{
Get the number of mutations in a table collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$num_mutations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A signed 64 bit integer \code{bit64::integer64}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$num_mutations()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-mutation_table_add_row"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-mutation_table_add_row}{}}}
\subsection{Method \code{mutation_table_add_row()}}{
Add a row to the mutations table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$mutation_table_add_row(
  site,
  node,
  parent = -1L,
  time = NaN,
  derived_state = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{site}}{integer scalar site row ID (0-based).}

\item{\code{node}}{integer scalar node row ID (0-based).}

\item{\code{parent}}{integer scalar parent mutation row ID (0-based, or \code{-1}).}

\item{\code{time}}{numeric scalar mutation time; use \code{NaN} for
\code{TSK_UNKNOWN_TIME}.}

\item{\code{derived_state}}{for the new mutation; accepts \code{NULL},
a raw vector, or a character of length 1.}

\item{\code{metadata}}{for the new mutation; accepts \code{NULL},
a raw vector, or a character of length 1.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See the \code{tskit Python} equivalent at
  \url{https://tskit.dev/tskit/docs/stable/python-api.html#tskit.MutationTable.add_row}.
  The function casts inputs to the expected class. \code{site},
  \code{node}, and \code{parent} must be non-\code{NA} integer scalars.
  \code{time} must be a numeric scalar that is finite or \code{NaN}
  (unknown time).
}

\subsection{Returns}{
Integer row ID (0-based) of the newly added mutation.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(ts_file)
n_before <- tc$num_mutations()
new_id <- tc$mutation_table_add_row(site = 0L, node = 0L, derived_state = "T")
new_id <- tc$mutation_table_add_row(
  site = 0L,
  node = 0L,
  parent = -1L,
  time = 1.5,
  derived_state = charToRaw("C"),
  metadata = "abc"
)
n_after <- tc$num_mutations()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-sequence_length"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-sequence_length}{}}}
\subsection{Method \code{sequence_length()}}{
Get the sequence length.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$sequence_length()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$sequence_length()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-time_units"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-time_units}{}}}
\subsection{Method \code{time_units()}}{
Get the time units string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$time_units()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$time_units()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-has_index"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-has_index}{}}}
\subsection{Method \code{has_index()}}{
Get whether the table collection has edge indexes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$has_index()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A logical.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$has_index()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-build_index"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-build_index}{}}}
\subsection{Method \code{build_index()}}{
Build edge indexes for this table collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$build_index()}\if{html}{\out{</div>}}
}

\subsection{Details}{
See the \code{tskit Python} equivalent at
  \url{https://tskit.dev/tskit/docs/latest/python-api.html#tskit.TableCollection.build_index}.
}

\subsection{Returns}{
No return value; called for side effects.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$has_index()
tc$drop_index()
tc$has_index()
tc$build_index()
tc$has_index()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-drop_index"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-drop_index}{}}}
\subsection{Method \code{drop_index()}}{
Drop edge indexes for this table collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$drop_index()}\if{html}{\out{</div>}}
}

\subsection{Details}{
See the \code{tskit Python} equivalent at
  \url{https://tskit.dev/tskit/docs/latest/python-api.html#tskit.TableCollection.drop_index}.
}

\subsection{Returns}{
No return value; called for side effects.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$has_index()
tc$drop_index()
tc$has_index()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-has_reference_sequence"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-has_reference_sequence}{}}}
\subsection{Method \code{has_reference_sequence()}}{
Get whether the table collection has a reference genome sequence.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$has_reference_sequence()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A logical.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file1 <- system.file("examples/test.trees", package = "RcppTskit")
tc_file2 <- system.file("examples/test_with_ref_seq.trees", package = "RcppTskit")
tc1 <- tc_load(tc_file1)
tc1$has_reference_sequence()
tc2 <- tc_load(tc_file2)
tc2$has_reference_sequence()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-file_uuid"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-file_uuid}{}}}
\subsection{Method \code{file_uuid()}}{
Get the UUID string of the file the table collection was
  loaded from.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$file_uuid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character; \code{NA_character_} when file is information is
  unavailable.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tc_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(tc_file)
tc$file_uuid()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-r_to_py"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-r_to_py}{}}}
\subsection{Method \code{r_to_py()}}{
This function saves a table collection from \code{R} to disk
  and loads it into reticulate \code{Python} for use with the
  \code{tskit Python} API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$r_to_py(tskit_module = get_tskit_py(), cleanup = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tskit_module}}{reticulate \code{Python} module of \code{tskit}.
By default, it calls \code{\link{get_tskit_py}} to obtain the module.}

\item{\code{cleanup}}{logical; delete the temporary file at the end of the function?}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \url{https://tskit.dev/tutorials/tables_and_editing.html#tables-and-editing}
  on what you can do with the tables.
}

\subsection{Returns}{
\code{TableCollection} object in reticulate \code{Python}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  ts_file <- system.file("examples/test.trees", package = "RcppTskit")
  tc_r <- tc_load(ts_file)
  is(tc_r)
  tc_r$print()

  # Transfer the table collection to reticulate Python and use tskit Python API
  tskit <- get_tskit_py()
  if (check_tskit_py(tskit)) {
    tc_py <- tc_r$r_to_py()
    is(tc_py)
    tmp <- tc_py$simplify(samples = c(0L, 1L, 2L, 3L))
    tmp
    tc_py$individuals$num_rows # 2
    tc_py$nodes$num_rows # 8
    tc_py$nodes$time # 0.0 ... 5.0093910
  }
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-print"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of a table collection and its contents.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list with two data.frames; the first contains table collection
  properties and their values; the second contains the number of rows in
  each table and the length of their metadata. All columns are characters
  since output types differ across the entries. Use individual getters
  to obtain raw values before they are converted to character.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ts_file <- system.file("examples/test.trees", package = "RcppTskit")
tc <- tc_load(file = ts_file)
tc$print()
tc
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCollection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TableCollection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
