% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-TreeSequence.R
\name{TreeSequence}
\alias{TreeSequence}
\title{Succinct Tree Sequence R6 Class}
\description{
An R6 class to hold an external pointer to a tskit tree sequence.
As an R6 class, it's functions will look Pythonic and hence resemble the
tskit Python API. Since the class only holds the pointer, it is lightweight.
The pointer does not provide direct view for users and currently there is
only a limited number of methods available to summarise the tree sequence.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pointer}}{external pointer to the tree sequence}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TreeSequence-new}{\code{TreeSequence$new()}}
\item \href{#method-TreeSequence-dump}{\code{TreeSequence$dump()}}
\item \href{#method-TreeSequence-write}{\code{TreeSequence$write()}}
\item \href{#method-TreeSequence-print}{\code{TreeSequence$print()}}
\item \href{#method-TreeSequence-summary}{\code{TreeSequence$summary()}}
\item \href{#method-TreeSequence-num_provenances}{\code{TreeSequence$num_provenances()}}
\item \href{#method-TreeSequence-num_populations}{\code{TreeSequence$num_populations()}}
\item \href{#method-TreeSequence-num_migrations}{\code{TreeSequence$num_migrations()}}
\item \href{#method-TreeSequence-num_individuals}{\code{TreeSequence$num_individuals()}}
\item \href{#method-TreeSequence-num_samples}{\code{TreeSequence$num_samples()}}
\item \href{#method-TreeSequence-num_nodes}{\code{TreeSequence$num_nodes()}}
\item \href{#method-TreeSequence-num_edges}{\code{TreeSequence$num_edges()}}
\item \href{#method-TreeSequence-num_trees}{\code{TreeSequence$num_trees()}}
\item \href{#method-TreeSequence-num_sites}{\code{TreeSequence$num_sites()}}
\item \href{#method-TreeSequence-num_mutations}{\code{TreeSequence$num_mutations()}}
\item \href{#method-TreeSequence-sequence_length}{\code{TreeSequence$sequence_length()}}
\item \href{#method-TreeSequence-time_units}{\code{TreeSequence$time_units()}}
\item \href{#method-TreeSequence-metadata_length}{\code{TreeSequence$metadata_length()}}
\item \href{#method-TreeSequence-r_to_py}{\code{TreeSequence$r_to_py()}}
\item \href{#method-TreeSequence-clone}{\code{TreeSequence$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-new"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{\link{TreeSequence}} from a file or a pointer.
  See \code{\link{ts_load}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$new(file, options = 0L, pointer = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{see \code{\link{ts_load}()}}

\item{\code{options}}{see \code{\link{ts_load}()}}

\item{\code{pointer}}{an external pointer to a tree sequence}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-dump"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-dump}{}}}
\subsection{Method \code{dump()}}{
Write a tree sequence to a file.
  See \code{\link{ts_dump}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$dump(file, options = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{see \code{\link{ts_dump}()}}

\item{\code{options}}{see \code{\link{ts_dump}()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-write"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-write}{}}}
\subsection{Method \code{write()}}{
Write a tree sequence to a file.
  Alias for \code{TreeSequence$dump(file, options)}.
  See \code{\link{ts_dump}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$write(file, options = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{see \code{\link{ts_dump}()}}

\item{\code{options}}{see \code{\link{ts_dump}()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-print"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of a tree sequence and its contents.
  See \code{\link{ts_print}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-summary"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-summary}{}}}
\subsection{Method \code{summary()}}{
Summary of properties and number of records in a tree sequence.
  See \code{\link{ts_summary}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-num_provenances"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-num_provenances}{}}}
\subsection{Method \code{num_provenances()}}{
Get the number of provenances in a tree sequence.
  See \code{\link{ts_summary}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$num_provenances()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-num_populations"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-num_populations}{}}}
\subsection{Method \code{num_populations()}}{
Get the number of populations in a tree sequence.
  See \code{\link{ts_summary}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$num_populations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-num_migrations"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-num_migrations}{}}}
\subsection{Method \code{num_migrations()}}{
Get the number of migrations in a tree sequence.
  See \code{\link{ts_summary}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$num_migrations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-num_individuals"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-num_individuals}{}}}
\subsection{Method \code{num_individuals()}}{
Get the number of individuals in a tree sequence.
  See \code{\link{ts_summary}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$num_individuals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-num_samples"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-num_samples}{}}}
\subsection{Method \code{num_samples()}}{
Get the number of samples (of nodes) in a tree sequence.
  See \code{\link{ts_summary}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$num_samples()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-num_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-num_nodes}{}}}
\subsection{Method \code{num_nodes()}}{
Get the number of nodes in a tree sequence.
  See \code{\link{ts_summary}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$num_nodes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-num_edges"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-num_edges}{}}}
\subsection{Method \code{num_edges()}}{
Get the number of edges in a tree sequence.
  See \code{\link{ts_summary}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$num_edges()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-num_trees"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-num_trees}{}}}
\subsection{Method \code{num_trees()}}{
Get the number of trees in a tree sequence.
  See \code{\link{ts_summary}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$num_trees()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-num_sites"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-num_sites}{}}}
\subsection{Method \code{num_sites()}}{
Get the number of sites in a tree sequence.
  See \code{\link{ts_summary}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$num_sites()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-num_mutations"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-num_mutations}{}}}
\subsection{Method \code{num_mutations()}}{
Get the number of mutations in a tree sequence.
  See \code{\link{ts_summary}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$num_mutations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-sequence_length"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-sequence_length}{}}}
\subsection{Method \code{sequence_length()}}{
Get the sequence length.
  See \code{\link{ts_summary}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$sequence_length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-time_units"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-time_units}{}}}
\subsection{Method \code{time_units()}}{
Get the time units string.
  See \code{\link{ts_summary}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$time_units()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-metadata_length"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-metadata_length}{}}}
\subsection{Method \code{metadata_length()}}{
Get the length of metadata in a tree sequence and its tables.
  See \code{\link{ts_metadata_length}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$metadata_length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-r_to_py"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-r_to_py}{}}}
\subsection{Method \code{r_to_py()}}{
Transfer a tree sequence from R to reticulate Python.
  See \code{\link{ts_r_to_py}()} for details and examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$r_to_py(tskit_module = get_tskit_py(), cleanup = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tskit_module}}{see \code{\link{ts_r_to_py}()}}

\item{\code{cleanup}}{see \code{\link{ts_r_to_py}()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeSequence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TreeSequence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeSequence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
