% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppTskit.R
\name{ts_load}
\alias{ts_load}
\alias{ts_read}
\title{Load a tree sequence from a file}
\usage{
ts_load(file, options = 0L)

ts_read(file, options = 0L)
}
\arguments{
\item{file}{a string specifying the full path to a tree sequence file.}

\item{options}{integer bitwise options (see details at
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_load}).}
}
\value{
A \code{\link{TreeSequence}} object.
}
\description{
Load a tree sequence from a file
}
\section{Functions}{
\itemize{
\item \code{ts_read()}: Alias for \code{ts_load()}

}}
\examples{
ts_file <- system.file("examples/test.trees", package = "RcppTskit")
ts <- ts_load(ts_file)
is(ts)
ts
ts$num_nodes()
# Also
ts <- TreeSequence$new(file = ts_file)
is(ts)
}
\seealso{
\code{\link[=TreeSequence]{TreeSequence$new}}
}
